HW06 code review fixes.
