+++ VAL.HW06.Task02.DevicesChain/Program.cs
@@ -6,99 +6,6 @@ using System.Threading.Tasks;
 
 namespace VAL.HW06.Task02.DevicesChain
 {
-
-    internal class Variables
-    {
-        public static int Id = 0;
-    }
-
-
-    internal class ElectricDevice
-    {
-        public int Id { get; private set; }
-        public ElectricDevice NextDevice { get; set; }
-
-        public ElectricDevice()
-        {
-            this.Id = Variables.Id++;
-        }
-
-        public bool PlugDevice(ConsumptionDevice device)
-        {
-            Console.Write($"Plugging device {device.Id}({device.ConsumptionPower}W) to {this.Id}({AvailablePower}W)... ");
-            if (this.NextDevice != null || AvailablePower < device.ConsumptionPower)
-            {
-                Console.WriteLine($"Failed");
-                return false;
-            }
-            this.NextDevice = device;
-            device.PreviousDevice = this;
-            Console.WriteLine($"Ok");
-            return true;
-        }
-
-        public virtual bool UnplugDevice()
-        {
-            if (this.NextDevice == null)
-            {
-                return false;
-            }
-            Console.WriteLine($"Unplugging device {this.NextDevice.Id.ToString()} from {this.Id}... Ok");
-            this.NextDevice.UnplugDevice();
-            this.NextDevice = null;
-            return true;
-        }
-
-        public virtual int AvailablePower { get; }
-
-        public string ChainList
-        {
-            get
-            {
-                if (this.NextDevice == null)
-                {
-                    return this.Id.ToString();
-                }
-                return this.Id.ToString() + " - " + this.NextDevice.ChainList;
-            }
-        }
-    }
-
-
-    internal class Generator : ElectricDevice
-    {
-        public int GeneratedPower { get; set; }
-
-        public override int AvailablePower
-        {
-            get
-            {
-                return this.GeneratedPower;
-            }
-        }
-    }
-
-
-    internal class ConsumptionDevice : ElectricDevice
-    {
-        public int ConsumptionPower { get; set; }
-        public ElectricDevice PreviousDevice { get; set; }
-
-        public override int AvailablePower
-        {
-            get
-            {
-                if (PreviousDevice == null)
-                {
-                    return 0;
-                }
-                return PreviousDevice.AvailablePower - this.ConsumptionPower;
-            }
-        }
-
-    }
-
-
     class Program
     {
         static void Main(string[] args)
