+++ VAL.HW06.Task02.DevicesChain/Variables.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace VAL.HW06.Task02.DevicesChain
+{
+    internal class Variables
+    {
+        public static int Id = 0;
+    }
+}
